+++ src/view/button-more.js
@@ -1,0 +1,6 @@
+const createButtonMoreMarkup = () => {
+  return(`<button class="films-list__show-more">Show more</button>`);
+}
+
+
+export {createButtonMoreMarkup}
