+++ src/view/film-container.js
@@ -1,0 +1,14 @@
+const createFilmContainerMarkup = () => {
+  return(`
+    <section class="films">
+      <section class="films-list">
+        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+        <div class="films-list__container">
+        </div>
+      </section>
+    </section>
+    `);
+}
+
+
+export {createFilmContainerMarkup}
