+++ src/view/film-most-commented.js
@@ -1,0 +1,12 @@
+const createFilmMostCommentedMarkup = () => {
+  return(`
+    <section class="films-list--extra">
+      <h2 class="films-list__title">Most commented</h2>
+      <div class="films-list__container">
+      </div>
+    </section>
+    `);
+}
+
+
+export {createFilmMostCommentedMarkup}
