+++ src/view/footer-stat.js
@@ -1,0 +1,6 @@
+const createFooterStatMarkup = () => {
+  return(`<p>130 291 movies inside</p>`);
+}
+
+
+export {createFooterStatMarkup}
